+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 ## 1.0.39
 * Fixed + Improved EMI compatibility (thanks to @PrototypeTrousers)
 * Fixed Number Configurator doesn't support Long
+* Allow separated bg for player inventory
 
 ## 1.0.38.d
 * Fixed editor resource rename doesn't work